+++ packages/components-js/tests/e2e/specs/checkbox.e2e.ts
@@ -11,7 +11,6 @@ import {
   setContentWithDesignSystem,
   setProperty,
   skipInBrowsers,
-  waitForInputTransition,
   waitForStencilLifecycle,
 } from '../helpers';
 import type { CheckboxState } from '@porsche-design-system/components';
@@ -27,6 +26,18 @@ const setChecked = async (locator: Locator, value: boolean) => {
   await setProperty(locator, 'checked', value);
 };
 
+async function performBoundaryClicks(host: Locator, page: Page) {
+  const coords = await host.boundingBox();
+
+  await page.mouse.click(coords.x + 1, coords.y + 1); // Top left corner
+  await page.mouse.click(coords.x + 1, coords.y + coords.height - 1); // Bottom left corner
+  await page.mouse.click(coords.x + coords.width - 1, coords.y + 1); // Top right corner
+  await page.mouse.click(coords.x + coords.width - 1, coords.y + coords.height - 1); // Bottom right corner
+  await page.mouse.click(coords.x + 1, coords.y + coords.height / 2); // Left center
+  await page.mouse.click(coords.x + coords.width - 1, coords.y + coords.height / 2); // Right center
+  await page.mouse.click(coords.x + coords.width / 2, coords.y + coords.height / 2); // Center center
+}
+
 const getBackgroundImage = (input: Locator) => getElementStyle(input, 'backgroundImage');
 const backgroundURL = 'url("data:image';
 
@@ -36,6 +47,8 @@ type InitOptions = {
   value?: string;
   label?: string;
   checked?: boolean;
+  indeterminate?: boolean;
+  disabled?: boolean;
   useSlottedLabel?: boolean;
   useSlottedMessage?: boolean;
   isWithinForm?: true;
@@ -52,6 +65,8 @@ const initCheckbox = (page: Page, opts?: InitOptions): Promise<void> => {
     value = '',
     form = '',
     checked = false,
+    indeterminate = false,
+    disabled = false,
     useSlottedLabel = false,
     useSlottedMessage = false,
     isWithinForm = false,
@@ -76,6 +91,8 @@ const initCheckbox = (page: Page, opts?: InitOptions): Promise<void> => {
     form && `form="${form}"`,
     loading && 'loading="true"',
     checked && 'checked="true"',
+    disabled && 'disabled="true"',
+    indeterminate && 'indeterminate="true"',
   ]
     .filter(Boolean)
     .join(' ');
@@ -126,7 +143,6 @@ test('should toggle checkbox when input is clicked', async ({ page }) => {
 
   // ensure that checked and indeterminate use different images
   await setProperty(host, 'indeterminate', true);
-  await waitForInputTransition(page);
   expect(checkedImage).not.toBe(await getBackgroundImage(input));
 });
 
@@ -134,68 +150,115 @@ test('should not toggle checkbox on click in loading state', async ({ page }) =>
   await initCheckbox(page, { loading: true });
   const host = getHost(page);
   const input = getInput(page);
-  await addEventListener(host, 'click');
-  await addEventListener(input, 'change');
+  await addEventListener(host, 'update');
 
+  await expect(host).toHaveJSProperty('checked', false);
   await input.click({ force: true });
-  const coords = await host.boundingBox();
-  await page.mouse.click(coords.x + 1, coords.y + 1); // click the top left corner
-  await page.mouse.click(coords.x + 1, coords.y + coords.height - 1); // click the bottom left corner
-  await page.mouse.click(coords.x + coords.width - 1, coords.y + 1); // click the top right corner
-  await page.mouse.click(coords.x + coords.width - 1, coords.y + coords.height - 1); // click the bottom right corner
-  await page.mouse.click(coords.x + 1, coords.y + coords.height / 2); // click the left center
-  await page.mouse.click(coords.x + coords.width - 1, coords.y + coords.height / 2); // click the right center
-  await page.mouse.click(coords.x + coords.width / 2, coords.y + coords.height / 2); // click the center center
+  await expect(host).toHaveJSProperty('checked', false);
 
-  expect((await getEventSummary(host, 'click')).counter).toBe(8);
-  expect((await getEventSummary(input, 'change')).counter).toBe(0);
+  await performBoundaryClicks(host, page);
+
+  expect((await getEventSummary(host, 'update')).counter).toBe(0);
 
   await setProperty(host, 'loading', false);
   await waitForStencilLifecycle(page);
 
   await input.click();
-  expect((await getEventSummary(host, 'click')).counter).toBe(9);
-  expect((await getEventSummary(input, 'change')).counter).toBe(1);
+  await expect(host).toHaveJSProperty('checked', true);
+
+  expect((await getEventSummary(host, 'update')).counter).toBe(1);
 });
 
-test('should not toggle checkbox when pressed space in focus in loading state', async ({ page }) => {
-  await initCheckbox(page, { loading: true });
+test('should not toggle checkbox on click in disabled state', async ({ page }) => {
+  await initCheckbox(page, { disabled: true });
   const host = getHost(page);
   const input = getInput(page);
-  await addEventListener(input, 'change');
+  await addEventListener(host, 'update');
 
-  await input.focus();
-  expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
+  await expect(host).toHaveJSProperty('checked', false);
+  await input.click({ force: true });
+  await expect(host).toHaveJSProperty('checked', false);
 
-  await page.keyboard.press('Space');
-  expect((await getEventSummary(input, 'change')).counter).toBe(0);
+  await performBoundaryClicks(host, page);
 
-  await setProperty(host, 'loading', false);
+  expect((await getEventSummary(host, 'update')).counter).toBe(0);
+
+  await setProperty(host, 'disabled', false);
   await waitForStencilLifecycle(page);
 
-  await page.keyboard.press('Space');
-  expect((await getEventSummary(input, 'change')).counter).toBe(1);
+  await input.click();
+  await expect(host).toHaveJSProperty('checked', true);
+
+  expect((await getEventSummary(host, 'update')).counter).toBe(1);
 });
 
-skipInBrowsers(['firefox', 'webkit'], () => {
-  test('should keep focus if state switches to loading', async ({ page }) => {
+test.describe('focus', () => {
+  test('should receive focus', async ({ page }) => {
     await initCheckbox(page);
     const input = getInput(page);
+
+    await input.click();
+    expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
+  });
+
+  test('should receive focus in loading state', async ({ page }) => {
+    await initCheckbox(page, { loading: true });
+    const input = getInput(page);
+
+    await input.focus();
+    expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
+    await expect(input).toHaveCSS('outline', 'rgb(26, 68, 234) solid 2px');
+    await expect(input).toHaveCSS('outline-offset', '2px');
+  });
+
+  test('should receive focus when focus is set programmatically', async ({ page }) => {
+    await initCheckbox(page);
+    const input = getInput(page);
+
+    await input.focus();
+    expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
+    await expect(input).toHaveCSS('outline', 'rgb(26, 68, 234) solid 2px');
+    await expect(input).toHaveCSS('outline-offset', '2px');
+  });
+
+  test('should not toggle checkbox when pressed space in focus in loading state', async ({ page }) => {
+    await initCheckbox(page, { loading: true });
     const host = getHost(page);
+    const input = getInput(page);
+    await addEventListener(input, 'change');
 
-    expect(await hasFocus(host)).toBe(false);
-    await page.keyboard.press('Tab');
-    expect(await hasFocus(host), 'after Tab').toBe(true);
+    await input.focus();
+    expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
 
-    await setProperty(host, 'loading', true);
-    await waitForStencilLifecycle(page);
-
-    expect(await hasFocus(host), 'focus when loading').toBe(true);
+    await page.keyboard.press('Space');
+    expect((await getEventSummary(input, 'change')).counter).toBe(0);
 
     await setProperty(host, 'loading', false);
     await waitForStencilLifecycle(page);
 
-    expect(await hasFocus(host), 'final focus').toBe(true);
+    await page.keyboard.press('Space');
+    expect((await getEventSummary(input, 'change')).counter).toBe(1);
+  });
+
+  skipInBrowsers(['firefox', 'webkit'], () => {
+    test('should keep focus if state switches to loading', async ({ page }) => {
+      await initCheckbox(page);
+      const host = getHost(page);
+
+      expect(await hasFocus(host)).toBe(false);
+      await page.keyboard.press('Tab');
+      expect(await hasFocus(host), 'after Tab').toBe(true);
+
+      await setProperty(host, 'loading', true);
+      await waitForStencilLifecycle(page);
+
+      expect(await hasFocus(host), 'focus when loading').toBe(true);
+
+      await setProperty(host, 'loading', false);
+      await waitForStencilLifecycle(page);
+
+      expect(await hasFocus(host), 'final focus').toBe(true);
+    });
   });
 });
 
@@ -221,19 +284,6 @@ test('should toggle checkbox when label text is clicked and not set input as act
   expect(await getActiveElementTagName(page)).toBe('P-CHECKBOX');
 });
 
-test('should check/uncheck checkbox when checkbox attribute is changed programmatically', async ({ page }) => {
-  await initCheckbox(page);
-  const input = getInput(page);
-
-  expect(await getBackgroundImage(input)).toBe('none');
-
-  await setProperty(input, 'checked', true);
-  expect(await getBackgroundImage(input)).toContain(backgroundURL);
-
-  await setProperty(input, 'checked', false);
-  expect(await getBackgroundImage(input)).toBe('none');
-});
-
 test('should check/uncheck checkbox when checkbox property is changed programmatically', async ({ page }) => {
   await initCheckbox(page);
   const input = getInput(page);
@@ -254,34 +304,29 @@ skipInBrowsers(['firefox', 'webkit'], () => {
     const input = getInput(page);
     const wrapper = getWrapper(page);
 
-    const getWrapperCursor = () => getElementStyle(wrapper, 'cursor');
-    const getInputCursor = () => getElementStyle(input, 'cursor');
-    const getInputPointerEvents = () => getElementStyle(input, 'pointerEvents');
-
-    expect(await getWrapperCursor()).toBe('auto');
-    expect(await getInputCursor()).toBe('pointer');
-    expect(await getInputPointerEvents()).toBe('auto');
+    await expect(wrapper).toHaveCSS('cursor', 'auto');
+    await expect(input).toHaveCSS('cursor', 'pointer');
+    await expect(input).toHaveCSS('pointer-events', 'auto');
 
     await setProperty(host, 'disabled', true);
-    await waitForInputTransition(page);
 
-    expect(await getWrapperCursor()).toBe('not-allowed');
-    expect(await getInputCursor()).toBe('default');
-    expect(await getInputPointerEvents()).toBe('none'); // prevents checkbox from being toggleable in disabled and especially loading state
+    await expect(wrapper).toHaveCSS('cursor', 'not-allowed');
+    await expect(input).toHaveCSS('cursor', 'default');
+    await expect(input).toHaveCSS('pointer-events', 'none'); // prevents checkbox from being toggleable in disabled and especially loading state
 
     await setProperty(host, 'disabled', false);
-    await waitForInputTransition(page);
 
-    expect(await getWrapperCursor()).toBe('auto');
-    expect(await getInputCursor()).toBe('pointer');
-    expect(await getInputPointerEvents()).toBe('auto');
+    await expect(wrapper).toHaveCSS('cursor', 'auto');
+    await expect(input).toHaveCSS('cursor', 'pointer');
+
+    await expect(input).toHaveCSS('pointer-events', 'auto');
 
     await setProperty(host, 'loading', true);
-    await waitForInputTransition(page);
 
-    expect(await getWrapperCursor()).toBe('not-allowed');
-    expect(await getInputCursor()).toBe('default');
-    expect(await getInputPointerEvents()).toBe('none'); // prevents checkbox from being toggleable in disabled and especially loading state
+    await expect(wrapper).toHaveCSS('cursor', 'not-allowed');
+    await expect(input).toHaveCSS('cursor', 'default');
+
+    await expect(input).toHaveCSS('pointer-events', 'none'); // prevents checkbox from being toggleable in disabled and especially loading state
   });
 });
 
@@ -296,12 +341,10 @@ test.describe('indeterminate state', () => {
     expect(await getBackgroundImage(input)).toBe('none');
 
     await setProperty(host, 'indeterminate', true);
-    await waitForInputTransition(page);
 
     expect(await getBackgroundImage(input)).toContain(backgroundURL);
 
     await setProperty(host, 'indeterminate', false);
-    await waitForInputTransition(page);
     expect(await getBackgroundImage(input)).toBe('none');
   });
 
@@ -311,7 +354,6 @@ test.describe('indeterminate state', () => {
     const input = getInput(page);
 
     await setProperty(host, 'indeterminate', true);
-    await waitForInputTransition(page);
     const indeterminateImage = await getBackgroundImage(input);
     expect(indeterminateImage, 'first indeterminate set').toContain(backgroundURL);
 
@@ -322,7 +364,6 @@ test.describe('indeterminate state', () => {
     expect(indeterminateImage).not.toBe(checkedImage);
 
     await setProperty(host, 'indeterminate', true);
-    await waitForInputTransition(page);
     expect(await getBackgroundImage(input), 'second indeterminate set').toContain(backgroundURL);
 
     await input.click();
@@ -335,7 +376,6 @@ test.describe('indeterminate state', () => {
     const input = getInput(page);
 
     await setProperty(host, 'indeterminate', true);
-    await waitForInputTransition(page);
     expect(await getBackgroundImage(input)).toContain(backgroundURL);
 
     await setChecked(input, true);
@@ -366,10 +406,10 @@ test.describe('lifecycle', () => {
     await waitForStencilLifecycle(page);
     const status = await getLifecycleStatus(page);
 
-    expect(status.componentDidUpdate['p-checkbox'], 'componentDidUpdate: p-checkbox').toBe(0);
+    expect(status.componentDidUpdate['p-checkbox'], 'componentDidUpdate: p-checkbox').toBe(1);
 
     expect(status.componentDidLoad.all, 'componentDidLoad: all').toBe(2);
-    expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(0);
+    expect(status.componentDidUpdate.all, 'componentDidUpdate: all').toBe(1);
   });
 });
 
@@ -393,6 +433,72 @@ test.describe('form', () => {
 
     expect((await getEventSummary(form, 'submit')).counter).toBe(1);
     expect(await getFormDataValue(form, name)).toBe(value);
+  });
+
+  test('should include name & value in FormData submit when checked in indeterminate state', async ({ page }) => {
+    const name = 'name';
+    const value = 'Hallo';
+    await initCheckbox(page, {
+      indeterminate: true,
+      name,
+      value,
+      checked: true,
+      isWithinForm: true,
+      markupAfter: '<button type="submit">Submit</button>',
+    });
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(value);
+  });
+
+  test('should not include name & value in FormData submit when disabled', async ({ page }) => {
+    const name = 'name';
+    const value = 'Hallo';
+    await initCheckbox(page, {
+      disabled: true,
+      name,
+      value,
+      checked: true,
+      isWithinForm: true,
+      markupAfter: '<button type="submit">Submit</button>',
+    });
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(null);
+  });
+
+  test.fixme('should not include name & value in FormData submit when loading', async ({ page }) => {
+    const name = 'name';
+    const value = 'Hallo';
+    await initCheckbox(page, {
+      loading: true,
+      name,
+      value,
+      checked: true,
+      isWithinForm: true,
+      markupAfter: '<button type="submit">Submit</button>',
+    });
+    const form = getForm(page);
+
+    await addEventListener(form, 'submit');
+    expect((await getEventSummary(form, 'submit')).counter).toBe(0);
+
+    await page.locator('button[type="submit"]').click();
+
+    expect((await getEventSummary(form, 'submit')).counter).toBe(1);
+    expect(await getFormDataValue(form, name)).toBe(null);
   });
 
   test('should only include name & value in FormData submit if checkbox is checked', async ({ page }) => {
@@ -483,5 +589,39 @@ test.describe('form', () => {
 
     expect((await getEventSummary(form, 'submit')).counter).toBe(1);
     expect(await getFormDataValue(form, name)).toBe(null);
+  });
+});
+
+test.describe('Event', () => {
+  skipInBrowsers(['firefox', 'webkit'], () => {
+    test('should trigger a update event when checkbox checked state has changed', async ({ page }) => {
+      const value = 'some-value';
+      await initCheckbox(page, { value });
+      const host = getHost(page);
+      const input = getInput(page);
+
+      await addEventListener(host, 'update');
+      expect((await getEventSummary(host, 'update')).counter).toBe(0);
+
+      await input.click();
+      await waitForStencilLifecycle(page);
+
+      expect((await getEventSummary(host, 'update')).counter).toBe(1);
+      expect((await getEventSummary(host, 'update')).details).toEqual([{ checked: true, name: 'some-name', value }]);
+    });
+    test('should trigger a blur event when the checkbox loses focus', async ({ page }) => {
+      await initCheckbox(page);
+      const host = getHost(page);
+      const input = getInput(page);
+
+      await addEventListener(host, 'blur');
+      expect((await getEventSummary(host, 'blur')).counter).toBe(0);
+
+      await input.click();
+      await input.press('Tab');
+      await waitForStencilLifecycle(page);
+
+      expect((await getEventSummary(host, 'blur')).counter).toBe(1);
+    });
   });
 });