+++ packages/components-react/src/examples/CheckboxExampleControlled.tsx
@@ -1,27 +1,26 @@
 import { useState } from 'react';
-import {
-  PCheckbox,
-  PText,
-  type CheckboxUpdateEventDetail,
-  type PCheckboxProps,
-} from '@porsche-design-system/components-react';
+import { PCheckbox, PText, type CheckboxUpdateEventDetail } from '@porsche-design-system/components-react';
 
 export const CheckboxControlledExamplePage = (): JSX.Element => {
-  const [value, setValue] = useState<PCheckboxProps['value']>('');
-
-  const onUpdate = (e: CustomEvent<CheckboxUpdateEventDetail>) => {
-    setValue(e.detail.checked ? e.detail.value : '');
+  const [state, setState] = useState({
+    'some-name': true,
+  });
+  const onUpdate = ({ detail: { name, checked } }: CustomEvent<CheckboxUpdateEventDetail>) => {
+    setState((prev) => ({
+      ...prev,
+      [name]: checked,
+    }));
   };
-
   return (
     <>
       <PCheckbox
         label="Some Label"
         name="some-name"
+        checked={state['some-name']}
         value="some-value"
-        onUpdate={(e) => onUpdate(e as CustomEvent<CheckboxUpdateEventDetail>)}
+        onUpdate={(e) => onUpdate(e)}
       />
-      <PText>Value: {value}</PText>
+      <PText>some-name: {state['some-name'].toString()}</PText>
     </>
   );
 };