+++ packages/components/src/components/checkbox/checkbox.spec.ts
@@ -2,7 +2,6 @@ import { Checkbox } from './checkbox';
 import * as applyConstructableStylesheetStylesUtils from '../../utils/applyConstructableStylesheetStyle';
 import { getSlottedAnchorStyles } from '../../styles';
 import { expect } from '@jest/globals';
-import { getCheckboxRadioButtonSafariRenderingFix } from '../../utils/form/applyCheckboxRadioButtonSafariRenderingFix';
 
 jest.mock('../../utils/dom');
 
@@ -26,6 +25,7 @@ const initComponent = (): Checkbox => {
   mockEmit = jest.fn();
 
   component.update = { emit: mockEmit } as any;
+  component.blur = { emit: mockEmit } as any;
   return component;
 };
 
@@ -38,11 +38,7 @@ describe('connectedCallback', () => {
     const component = initComponent();
 
     component.connectedCallback();
-    expect(applyConstructableStylesheetStylesSpy).toHaveBeenCalledWith(
-      component.host,
-      getSlottedAnchorStyles,
-      getCheckboxRadioButtonSafariRenderingFix
-    );
+    expect(applyConstructableStylesheetStylesSpy).toHaveBeenCalledWith(component.host, getSlottedAnchorStyles);
   });
 });
 
@@ -80,6 +76,22 @@ describe('componentDidLoad', () => {
     const setFormValueSpy = jest.spyOn(component['internals'], 'setFormValue' as any);
     component.componentDidLoad();
     expect(setFormValueSpy).not.toHaveBeenCalled();
+  });
+});
+
+describe('onBlur', () => {
+  it('should stop propagation and emit blur event on onBlur', () => {
+    const component = initComponent();
+    const event = {
+      stopPropagation: jest.fn(),
+      stopImmediatePropagation: jest.fn(),
+    } as unknown as Event;
+
+    component['onBlur'](event);
+
+    expect(event.stopPropagation).toHaveBeenCalled();
+    expect(event.stopImmediatePropagation).toHaveBeenCalled();
+    expect(mockEmit).toHaveBeenCalledWith(event);
   });
 });